Converted review panel from nested tables to list - cleaner code, better result; TODOs added
